git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi@5385 99afaf0d-6903-0410-885a-c66a8bbb5f81

+++ V2.x/trunk/SpagoBIProject/src/it/eng/spagobi/engines/chart/bo/charttypes/XYCharts/BlockChart.java
@@ -103,7 +103,7 @@ public class BlockChart extends XYCharts {
         chart.removeLegend();
         chart.setBackgroundPaint(Color.white);
         SymbolAxis scaleAxis = new SymbolAxis(null,new String[]{"1","2","3","4","5"});
-        scaleAxis.setRange(zvalues[1], new Double(zrangeMax));
+        scaleAxis.setRange(zvalues[1], new Double(zrangeMax).doubleValue());
         scaleAxis.setPlot(new PiePlot());
         scaleAxis.setGridBandsVisible(false);
         PaintScaleLegend psl = new PaintScaleLegend(paintScale, scaleAxis);
